<전화번호 관리 프로그램>
System.out.println("이름 : ");
String name = input.nextLine();

System.out.println("전화번호 : ");
String phoneNumber = input.nextLine();
위와 같이 했는데 이름 : 전화번호 : 가 같이 출력되어 이름을 입력할 수 없음....

+++ src/main/java/homework/HomeWork27.java
@@ -5,48 +5,140 @@ import java.util.Scanner;
 public class HomeWork27 {
     static Scanner input = new Scanner(System.in);
 
-    public static class PhoneInfo {
+    interface Menu{
+        int INPUT = 1, SEARCH = 2, DELETE = 3, END = 4;
+    }
+
+    interface InputSelect{
+        int BASIC = 1, UNIV = 2, COMPANY = 3;
+    }
+
+    public static class MenuInputException extends Exception{
+        int wrongChoice;
+
+        public MenuInputException(int choice){
+            wrongChoice = choice;
+        }
+        public void printException(){
+            System.out.println(wrongChoice + "에 해당하는 선택은 존재하지 않습니다.");
+        }
+    }
+
+    public static class PhoneInfo{
         String name;
         String phoneNumber;
-        String birthday;
 
-        public PhoneInfo(String name, String phoneNumber, String birthday) {
+        public PhoneInfo(String name, String num) {
             this.name = name;
-            this.phoneNumber = phoneNumber;
-            this.birthday = birthday;
-        }
-
-        public PhoneInfo(String name, String phoneNumber) {
-            this.name = name;
-            this.phoneNumber = phoneNumber;
+            phoneNumber = num;
         }
 
         public void printData(){
             System.out.println("이름 : " + name);
             System.out.println("전화번호 : " + phoneNumber);
-            if (birthday != null){
-                System.out.println("생일 : " + birthday);
-            }
+        }
+    }
+
+    public static class PhoneUnivInfo extends PhoneInfo{
+        String major;
+        int year;
+
+        public PhoneUnivInfo(String name, String num, String major, int year){
+            super(name, num);
+            this.major = major;
+            this.year = year;
+        }
+
+        public void printData(){
+            super.printData();
+            System.out.println("전공 : " + major);
+            System.out.println("학년 : " + year);
+        }
+    }
+
+    public static class PhoneCompanyInfo extends PhoneInfo{
+        String company;
+
+        public PhoneCompanyInfo(String name, String num, String company){
+            super(name, num);
+            this.company = company;
+        }
+
+        public void printData(){
+            super.printData();
+            System.out.println("회사 : " + company);
         }
     }
 
     public static class PhoneBookManager{
+        public static final int NOT_FOUND = -1;
         PhoneInfo[] phoneInfos = new PhoneInfo[100];
-        int infos = 0;
+        int count = 0;
 
-        public void input(){
-            System.out.println("데이터 입력을 시작합니다..");
-
+        private PhoneInfo basicInfo(){
             System.out.println("이름 : ");
             String name = input.nextLine();
 
             System.out.println("전화번호 : ");
             String phoneNumber = input.nextLine();
 
-            System.out.println("생일 : ");
-            String birthday = input.nextLine();
+            return new PhoneInfo(name, phoneNumber);
+        }
 
-            phoneInfos[infos++] = new PhoneInfo(name, phoneNumber, birthday);
+        private PhoneInfo univInfo(){
+            System.out.println("이름 : ");
+            String name = input.nextLine();
+
+            System.out.println("전화번호 : ");
+            String phoneNumber = input.nextLine();
+
+            System.out.println("전공 : ");
+            String major = input.nextLine();
+
+            System.out.println("학년 : ");
+            int year = input.nextInt();
+
+            return new PhoneUnivInfo(name, phoneNumber, major, year);
+        }
+
+        private PhoneInfo companyInfo(){
+            System.out.println("이름 : ");
+            String name = input.nextLine();
+
+            System.out.println("전화번호 : ");
+            String phoneNumber = input.nextLine();
+
+            System.out.println("회사 : ");
+            String company = input.nextLine();
+
+            return new PhoneCompanyInfo(name, phoneNumber, company);
+        }
+
+        public void input() throws MenuInputException{
+            System.out.println("데이터 입력을 시작합니다..");
+            System.out.println("1. 일반, 2. 대학 3. 회사");
+            System.out.println("선택 >> ");
+            int choice = input.nextInt();
+            PhoneInfo info = null;
+
+            if (choice < InputSelect.BASIC || choice > InputSelect.COMPANY){
+                MenuInputException menuInputException = new MenuInputException(choice);
+                throw menuInputException;
+            }
+
+                switch (choice){
+                    case InputSelect.BASIC:
+                        info = basicInfo();
+                        break;
+                    case InputSelect.UNIV:
+                        info = univInfo();
+                        break;
+                    case InputSelect.COMPANY:
+                        info = companyInfo();
+                        break;
+                }
+
+            phoneInfos[count++] = info;
             System.out.println("데이터 입력이 완료되었습니다.");
             System.out.println("");
         }
@@ -54,18 +146,46 @@ public class HomeWork27 {
         public void search(){
             System.out.println("데이터 검색을 시작합니다.");
 
-            System.out.printf("이름 : ");
-            String searchName = input.nextLine();
+            System.out.println("이름 : ");
+            String name = input.nextLine();
 
-            //compareTo 메소드에서 비교한 결과를 통해 해당 정보를 출력
-        }
+            int searchedIndex = getIndexOf(name);
 
-        private int compareTo(String name){
-            //search에서 입력한 이름과 phoneInfos[]에 입력한 이름과 비교
-
+            if (searchedIndex != NOT_FOUND){
+                phoneInfos[searchedIndex].printData();
+            }
+            System.out.println("데이터 검색이 완료되었습니다.");
         }
 
         public void delete(){
+            System.out.println("데이터 삭제를 시작합니다..");
+
+            System.out.println("이름 : ");
+            String name = input.nextLine();
+
+            int searchedIndex = getIndexOf(name);
+
+            PhoneInfo[] newPhoneInfos = new PhoneInfo[100];
+            int newIndex = 0;
+
+            for (int i = 0; i < phoneInfos.length; i++){
+                if (searchedIndex == i){
+                    continue;
+                }
+                newPhoneInfos[newIndex++] = phoneInfos[i];
+            }
+            phoneInfos = newPhoneInfos;
+            System.out.println("데이터 삭제가 완료되었습니다.");
+        }
+
+        private int getIndexOf(String name){
+            PhoneBookManager phoneBookManager = new PhoneBookManager();
+            for (int i = 0; i < count; i++){
+                if (phoneInfos[i].name == name){
+                    return i;
+                }
+            }
+            return -1;
         }
     }
 
@@ -88,22 +208,34 @@ public class HomeWork27 {
         PhoneBookManager phoneBookManager = new PhoneBookManager();
 
         while (true){
-            select.select();
-            choice = input.nextInt();
-            switch (choice){
-                case 1:
-                    phoneBookManager.input();
-                    break;
-                case 2:
-                    phoneBookManager.search();
-                    break;
-                case 3:
-                    phoneBookManager.delete();
-                    break;
-                case 4:
-                    System.out.println("프로그램을 종료합니다.");
-                    return;
+            try {
+                select.select();
+                choice = input.nextInt();
+
+                if (choice < Menu.INPUT || choice > Menu.END){
+                    MenuInputException menuInputException = new MenuInputException(choice);
+                    throw menuInputException;
+                }
+
+                switch (choice){
+                    case Menu.INPUT:
+                        phoneBookManager.input();
+                        break;
+                    case Menu.SEARCH:
+                        phoneBookManager.search();
+                        break;
+                    case Menu.DELETE:
+                        phoneBookManager.delete();
+                        break;
+                    case Menu.END:
+                        System.out.println("프로그램을 종료합니다.");
+                        return;
+                }
+            }catch (MenuInputException exception){
+                exception.printException();
+                System.out.println("메뉴 선택을 처음부터 다시 진행합니다.");
             }
+
         }
     }
 }
